+++ sharding-orchestration/sharding-orchestration-core/src/main/java/io/shardingsphere/orchestration/internal/state/node/StateNodeStatus.java
@@ -15,7 +15,7 @@
  * </p>
  */
 
-package io.shardingsphere.orchestration.internal.state;
+package io.shardingsphere.orchestration.internal.state.node;
 
 /**
  * State node status.
